remove dependency on tweeter class in printer
